+++ src/OmniSharp.Abstractions/MSBuild/Discovery/DiscoveryType.cs
@@ -2,7 +2,6 @@
 {
     public enum DiscoveryType
     {
-        StandAlone = 0,
         DeveloperConsole = 1,
         VisualStudioSetup = 2,
         Mono = 3,