+++ src/OmniSharp.Roslyn/Utilities/CodeActionOptionsFactory.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.CodeAnalysis.ExternalAccess.OmniSharp.CodeActions;
+using Microsoft.CodeAnalysis.ExternalAccess.OmniSharp.ImplementType;
+using OmniSharp.Options;
+
+namespace OmniSharp.Roslyn.CodeActions
+{
+    internal static class CodeActionOptionsFactory
+    {
+        public static OmniSharpCodeActionOptions Create(OmniSharpOptions options)
+            => new(new OmniSharpImplementTypeOptions(
+                (OmniSharpImplementTypeInsertionBehavior)options.ImplementTypeOptions.InsertionBehavior,
+                (OmniSharpImplementTypePropertyGenerationBehavior)options.ImplementTypeOptions.PropertyGenerationBehavior));
+    }
+}