+++ test-assets/test-projects/Net50Project/net50-lib/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Net50Lib
-{
-    public class Class1
-    {
-    }
-}