+++ tests/OmniSharp.MSBuild.Tests/AbstractMSBuildTestFixture.cs
@@ -4,8 +4,8 @@ using System.Composition.Hosting.Core;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using OmniSharp.Host.Services;
 using OmniSharp.MSBuild.Discovery;
+using OmniSharp.Roslyn.Utilities;
 using OmniSharp.Services;
 using TestUtility;
 using Xunit.Abstractions;
@@ -22,8 +22,16 @@ namespace OmniSharp.MSBuild.Tests
             : base(output)
         {
             _assemblyLoader = new AssemblyLoader(this.LoggerFactory);
-            _analyzerAssemblyLoader = new DefaultAnalyzerAssemblyLoader();
-            _msbuildLocator = MSBuildLocator.CreateStandAlone(this.LoggerFactory, _assemblyLoader);
+            _analyzerAssemblyLoader = ShadowCopyAnalyzerAssemblyLoader.Instance;
+
+            // Since we can only load MSBuild once into our process we need to include
+            // prerelease version so that our .NET 7 tests will pass.
+            var configuration = new Dictionary<string, string>
+            {
+                ["sdk:IncludePrereleases"] = bool.TrueString
+            }.ToConfiguration();
+
+            _msbuildLocator = MSBuildLocator.CreateDefault(this.LoggerFactory, _assemblyLoader, configuration);
 
             // Some tests require MSBuild to be discovered early
             // to ensure that the Microsoft.Build.* assemblies can be located