+++ tests/OmniSharp.Tests/LinePositionSpanTextChangeFacts.cs
@@ -26,9 +26,8 @@ namespace OmniSharp.Tests
 
                 var textChange = new TextChange(TextSpan.FromBounds(8, 11), "\n}");
 
-                var adjustedTextChanges = TextChanges.Convert(text, textChange);
+                var adjustedTextChange = TextChanges.Convert(text, textChange);
 
-                var adjustedTextChange = adjustedTextChanges.First();
                 Assert.Equal("\r\n}", adjustedTextChange.NewText);
                 Assert.Equal(0, adjustedTextChange.StartLine);
                 Assert.Equal(7, adjustedTextChange.StartColumn);
@@ -48,9 +47,8 @@ namespace OmniSharp.Tests
 
                 var textChange = new TextChange(TextSpan.FromBounds(5, 7), "\r\n {\r");
 
-                var adjustedTextChanges = TextChanges.Convert(text, textChange);
+                var adjustedTextChange = TextChanges.Convert(text, textChange);
 
-                var adjustedTextChange = adjustedTextChanges.First();
                 Assert.Equal("\r\n {\r\n", adjustedTextChange.NewText);
                 Assert.Equal(0, adjustedTextChange.StartLine);
                 Assert.Equal(5, adjustedTextChange.StartColumn);